test: move catalog-merge regression test into its own module
Keeps openrouter_tests.rs flat for the test-size ratchet.
