+++ crates/jcode-provider-openrouter-runtime/src/openrouter_tests.rs
@@ -10,9 +10,9 @@ use std::sync::mpsc;
 use std::time::Duration;
 use tempfile::TempDir;
 
-struct SharedEnvLock;
+pub(crate) struct SharedEnvLock;
 
-static ENV_LOCK: SharedEnvLock = SharedEnvLock;
+pub(crate) static ENV_LOCK: SharedEnvLock = SharedEnvLock;
 
 impl SharedEnvLock {
     /// Acquire the process-global test env lock.
@@ -22,24 +22,24 @@ impl SharedEnvLock {
     /// process env state, so a panic in one test must not cascade into a
     /// flood of unrelated `PoisonError` failures across every other test
     /// that takes this lock.
-    fn lock(&self) -> std::sync::MutexGuard<'static, ()> {
+    pub(crate) fn lock(&self) -> std::sync::MutexGuard<'static, ()> {
         jcode_base::storage::lock_test_env()
     }
 }
 
-struct EnvVarGuard {
+pub(crate) struct EnvVarGuard {
     key: &'static str,
     previous: Option<OsString>,
 }
 
 impl EnvVarGuard {
-    fn set(key: &'static str, value: impl AsRef<std::ffi::OsStr>) -> Self {
+    pub(crate) fn set(key: &'static str, value: impl AsRef<std::ffi::OsStr>) -> Self {
         let previous = std::env::var_os(key);
         jcode_base::env::set_var(key, value);
         Self { key, previous }
     }
 
-    fn remove(key: &'static str) -> Self {
+    pub(crate) fn remove(key: &'static str) -> Self {
         let previous = std::env::var_os(key);
         jcode_base::env::remove_var(key);
         Self { key, previous }
@@ -299,52 +299,6 @@ fn named_openai_compatible_provider_exposes_static_models_as_routes() {
             && route.api_method == "openai-compatible:comtegra-test"
             && route.available
     }));
-}
-
-#[test]
-fn named_profile_static_models_survive_live_catalog_refresh() {
-    let _lock = ENV_LOCK.lock();
-    let _namespace = EnvVarGuard::remove("JCODE_OPENROUTER_CACHE_NAMESPACE");
-    let _key = EnvVarGuard::set("TEST_NAMED_MERGE_KEY", "test-key");
-
-    let profile = jcode_base::config::NamedProviderConfig {
-        base_url: "https://llm.example.test/v1".to_string(),
-        api_key_env: Some("TEST_NAMED_MERGE_KEY".to_string()),
-        model_catalog: true,
-        default_model: Some("my-custom-model".to_string()),
-        models: vec![jcode_base::config::NamedProviderModelConfig {
-            id: "my-custom-model".to_string(),
-            ..Default::default()
-        }],
-        ..Default::default()
-    };
-
-    let provider = OpenRouterProvider::new_named_openai_compatible("user-compat", &profile)
-        .expect("named profile should initialize");
-    assert!(provider.is_user_named_profile());
-    assert!(provider.should_merge_static_models_with_live_catalog());
-
-    // Simulate a completed background `/models` catalog refresh that does not
-    // include the user's config-declared model.
-    {
-        let mut cache = provider.models_cache.blocking_write();
-        cache.models = vec![jcode_provider_openrouter::ModelInfo {
-            id: "vendor-live-model".to_string(),
-            name: "vendor live model".to_string(),
-            context_length: Some(128_000),
-            pricing: Default::default(),
-            created: None,
-        }];
-        cache.fetched = true;
-        cache.cached_at = Some(1);
-    }
-
-    let models = provider.available_models_display();
-    assert!(
-        models.iter().any(|m| m == "my-custom-model"),
-        "config-declared model should survive live catalog refresh: {models:?}"
-    );
-    assert!(models.iter().any(|m| m == "vendor-live-model"));
 }
 
 #[test]
